requirements.txt
decord==0.6.0
fvcore==0.1.5.post20221221
matplotlib==3.7.5
mmcv_full==1.5.0
moviepy==1.0.3
numpy==1.24.4
onnxruntime_gpu==1.17.0
opencv_contrib_python==4.11.0.86
opencv_python==4.10.0.84
openpyxl==3.1.5
pandas==2.0.3
Pillow==10.4.0
pymemcache==4.0.0
PySide6==6.6.3.1
PySide6_Addons==6.6.3.1
PySide6_Essentials==6.6.3.1
Requests==2.32.4
scipy==1.10.1
tqdm==4.66.5
ultralytics==8.3.55
